Fuck this is all a giant ass HACK ...
So now since page navigations do not case the browser to refresh,
the side effect is that navigating to a new page does not cause
the browser to reposition itself to the top of the new page.
So..., there's a million ways to do that, check out

https://stackoverflow.com/questions/36904185/react-router-scroll-to-top-on-every-transition

HOWEVER, none of the solutions listed are risk-free and 100% browser
compatible. In addition, this breaks the intra-page jump from the
'Explore' button inside the home page, but only on Safari ..
great eh...

The best bandage solution here is to just manually
invoke `window.scrollTo(0, 0)` inside of `componentDidMount()`
within `ContentPageSkeleton`, so this will only work for content pages,
which is the desired behavior.