Convert stylesheet links in index.html's head markup to 'preload' links
This patch takes a script from CRA's discussion on this topic (at the URL below)
and converts all stylesheet links in index.html's head markup to 'preload' links.
This is run as part of the "build" step action specified in 'package.json'.

Article on preloading resources:

https://web.dev/preload-critical-assets/

CRA issue and discussion:

https://github.com/facebook/create-react-app/issues/3319