Fix/issue #95 (#220)
This patch consists of an attempted solution for fixing the issue presented in Issue #95.

Note that the solution consists of both layout and style changes to the component `ImageSlidingGallery`. In addition to enforcing a style of `max-height: 320px` on the image within the component, the solution also changes the hierarchy of the navigation button container div of the gallery such that the nav buttons do not "jump" as much as would have seen previously.

The solution is not perfect, and "page jump" can still occur since the height of the caption text is variable and not enforced with a fixed bounded height. However, the rendering output with this change is overall still more palatable than would have been otherwise.

In addition, this change also consists of a change in one of the artworks used in the Future page, to ensure better consistency in dimension and aspect ration across all the image artwork used in one of the instances of this component, under the Future page.