Minor refinement in homepage explore gateway layout (#223)
This patch introduces minor changes to the layout of the Explore gateway within the home page. Specifically, it moves the Intro and Grid components of the Explore gateway to be adjacent, with the goal of making the navigation intent on the page more clear. The component that has the quote and "Get Involved" subsection has moved to under the Explore grid menu (i.e. refactored into its own component named `ExploreGatewayFooter`).