Refine content page local page navigation - Part 2 (#226)
This patch attempts to address the shortcoming of lacking an effective mechanism within a content page to navigate easily across the page when viewed under a mobile-like display size.

The solution is to introduce a nav menu bar component (i.e. `ContentPageBottomNavBar`) that always appears at the bottom of the screen at a fixed position and offset, and offers navigation links to jump to the top of the page, ToC of the content page, as well as support to navigate to the previous and next connected content pages.