[CRITICAL] Fix content page loading error
This patch attempts to address a critical issue where content pages
fail to load successfully on first load.

This issue seems to stem from the change made in #227 (b127792)
where the migration of project version string HTML comments to
a direct descendent of the root node seem to cause issues to React
when page loads to the next content page.

In addition, this patch also updates versionutil.py to be able to
update prerelease or build component of version string without
having to bump any of the major, minor, or patch components.

Bumped version to `0.4.3-alpha.4+08.11.21-0810`