Workaround for require.context (#249)
This patch introduces a workaround for the restricted usage of `require.context` in a Jest setting, by introducing a hand-crafted version of the "function" that mimics the behavior of `require.context` as far as the functionalities required by this project are concerned (i.e. dynamically loading images). In addition, conditionally apply either one of the two based on a predicate determine whether or not the current node environment is a test environment.

The benefit of this patch is that all unit and snapshot tests previously disabled are resurrected.

Updated project version to `0.5.3-alpha.1+08.21.21-1555`