[CRITICAL FIX] Fix crash on WebKit-based browsers - 08.25.21
This patch consists of a critical fix to a bug introduced by the
combination of the immediate previous two commits `03209b7` and `bc32f08`.
The bug is caused by the combination of the facts that an extra forward slash
was introduced in addPreloadWoff.js, but was disabled in `03209b7` and later
re-enabled in `bc32f08`, thus effectively did not surface in `03209b7`.
It appears that the extra forward slash introduced in addPreloadWoff.js confuses
the next immediate script addPreloadLinks.js. The fix is to simply remove this
extra forward slash introduced.