Fix/Issue 260
This patch fixes Issue #260 where the images within the Big Cats Spots illustration
component are vertically stretched.

The problem comes from the fact that the <img> element within the `TextRevealDetailed`
component had a 200:200 aspect ratio, but the underlying image assets used have
dimensions of 200w 160h, of aspect ratio 200:160.

The solution is to update the hardcoded aspect ratio on the <img> element to match
that of those image assets. This is not ideal but suffice for now since the
`TextRevealDetailed` component is only used within this component for the moment.