Fix image sliding modal instance dimension overflow issue under Ecology page (#291)
This patch fixes a bug in the Ecology page under the "Where Cheetahs Live" subsection, where the Namibia Biomes `ImageSlidingModal`'s overflows the outer containers when the width of the viewport is less than the width of the sliding modal (i.e. `480px`). This is particularly obvious when viewing on a mobile device with small display size (e.g., iPhone 13 mini).

The underlying cause was that the Namibia Biomes component was nested under `ContentPageSubsectionTwoColumnContentTemplate `, which apparently does not handle child column content dimension overflow very well. The temporary albeit effective solution is to use `FluidTwoColumnContainer` instead.